requirements.txt
# Core dependencies
numpy==2.4.2
scipy==1.17.0
Pillow==12.1.0
python-dotenv==1.2.1
pandas==3.0.0
packaging==25.0
PyYAML==6.0.3

# Video processing
ffmpeg-python==0.2.0
opencv-python==4.13.0.92
OpenTimelineIO==0.18.1
soundfile==0.13.1

# Machine Learning - PyTorch
torch==2.10.0
torchvision==0.25.0
safetensors==0.7.0

# Machine Learning - Transformers & Vision models
transformers==5.0.0
tokenizers==0.22.2
huggingface_hub==1.4.0
datasets==4.5.0
sentencepiece==0.2.1

# Machine Learning - MLX for Mac (Qwen2-VL)
# Note: MLX only works on Apple Silicon Macs
mlx==0.30.5
mlx-lm==0.30.6
mlx-metal==0.30.5
mlx-vlm==0.3.11

# LLM frameworks - LangChain
langchain==1.2.8
langchain-core==1.2.8
langchain-google-genai==4.2.0
langchain-ollama==1.0.1
langsmith==0.6.8

# LLM frameworks - LangGraph
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3

# LLM providers
openai==2.16.0
ollama==0.6.1
google-genai==1.62.0
google-auth==2.48.0

# Observability - Langfuse & OpenTelemetry
langfuse==3.12.1
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1

# Data validation & serialization
pydantic==2.12.5
pydantic_core==2.41.5
orjson==3.11.7
ormsgpack==1.12.2

# HTTP & networking
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.6.3
websockets==15.0.1
h11==0.16.0

# API frameworks
fastapi==0.128.1
starlette==0.50.0
uvicorn==0.40.0
anyio==4.12.1
sniffio==1.3.1

# Cryptography & security
cryptography==46.0.4
certifi==2026.1.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
rsa==4.9.1

# Data processing
pyarrow==23.0.0
dill==0.4.0
multiprocess==0.70.18
xxhash==3.6.0
zstandard==0.25.0

# File & storage
filelock==3.20.3
filetype==1.2.0
fsspec==2025.10.0
hf-xet==1.2.0

# CLI & utilities
click==8.3.1
typer-slim==0.21.1
shellingham==1.5.4
tqdm==4.67.3
backoff==2.2.1
tenacity==9.1.3

# Type annotations & validation
annotated-doc==0.0.4
annotated-types==0.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0

# Templating & text processing
Jinja2==3.1.6
MarkupSafe==3.0.3
regex==2026.1.15

# Math & symbolic computation
sympy==1.14.0
mpmath==1.3.0
networkx==3.6.1

# JSON & data manipulation
jsonpatch==1.33
jsonpointer==3.0.0
jiter==0.13.0

# Protocol buffers & API
protobuf==6.33.5
googleapis-common-protos==1.72.0

# Parsing & encoding
charset-normalizer==3.4.4
idna==3.11
frozenlist==1.8.0
multidict==6.7.1
propcache==0.4.1
yarl==1.22.0

# Utilities
attrs==25.4.0
six==1.17.0
python-dateutil==2.9.0.post0
wrapt==1.17.3
uuid_utils==0.14.0
distro==1.9.0
future==1.0.0

# System packages
cffi==2.0.0
pycparser==3.0
setuptools==80.10.2
importlib_metadata==8.7.1
zipp==3.23.0
